Move copyDependencies to common parameters and all server goals
