+++ liberty-maven-plugin/src/main/java/io/openliberty/tools/maven/server/CopyDependencies.java
@@ -13,16 +13,30 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package io.openliberty.tools.maven.applications;
+package io.openliberty.tools.maven.server;
 
 import java.util.ArrayList;
 import java.util.List;
 import org.apache.maven.plugins.annotations.Parameter;
 
 /**
- * A collection of Dependency with an optional location.
+ * A class to store a set of CopyDependencies.
  */
-public class DependencyGroup {
+public class CopyDependencies {
+
+    /**
+     * The location to copy dependencies to. This can be a full path, or a path relative to
+     * ${server.config.dir}. The default is ${server.config.dir}/lib/global.
+     */
+    @Parameter
+    private String location = "lib/global";
+
+    /**
+     * Boolean to indicate whether to strip versions from the file names when copying.
+     * The default is false.
+     */
+    @Parameter(defaultValue="false")
+    private Boolean stripVersion;
 
     /**
      * A list of Dependency to copy.
@@ -30,13 +44,10 @@ public class DependencyGroup {
     private List<Dependency> copyDependencyList = new ArrayList<Dependency>();
     
     /**
-     * Optional location to copy the Dependency to. This can be a full path, or a path relative to
-     * ${server.config.dir}. The location in the containing CopyDependencies configuration will be 
-     * used if nothing is specified here.
+     * A list of DependencyGroup to copy.
      */
-    @Parameter
-    private String location = null;
-
+    private List<DependencyGroup> copyDependencyGroupList = new ArrayList<DependencyGroup>();
+    
     public String getLocation() {
         return location;
     }
@@ -45,13 +56,12 @@ public class DependencyGroup {
         this.location = location;
     }
     
-    /**
-     * Boolean to indicate whether to strip versions from the file names when copying.
-     * The stripVersion in the containing CopyDependencies configuration will be 
-     * used if nothing is specified here.
-     */
-    @Parameter
-    private Boolean stripVersion = null;
+    public boolean isStripVersion() {
+        if (this.stripVersion == null) {
+            return false;
+        }
+        return this.stripVersion.booleanValue();
+    }
 
     public Boolean getStripVersion() {
         return this.stripVersion;
@@ -60,6 +70,7 @@ public class DependencyGroup {
     public void setStripVersion(boolean strip) {
         this.stripVersion = new Boolean(strip);
     }
+    
     /**
      * Get all the current Dependency to copy.
      *
@@ -76,6 +87,24 @@ public class DependencyGroup {
      */
     public void addDependency(Dependency dependency) {
         copyDependencyList.add(dependency);
+    }
+
+    /**
+     * Get all the current DependencyGroups to copy.
+     *
+     * @return A list with the DependencyGroup to copy.
+     */
+    public List<DependencyGroup> getDependencyGroups() {
+        return copyDependencyGroupList;
+    }
+
+    /**
+     * Add a DependencyGroup into a list.
+     *
+     * @param DependencyGroup
+     */
+    public void addDependencyGroup(DependencyGroup dependencyGroup) {
+        copyDependencyGroupList.add(dependencyGroup);
     }
 
 }
