+++ liberty-maven-plugin/src/main/java/io/openliberty/tools/maven/server/Dependency.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package io.openliberty.tools.maven.applications;
+package io.openliberty.tools.maven.server;
 
 import org.apache.maven.plugins.annotations.Parameter;
